adds label times folder
